Improve/extend the JS examples.
